feat(CoordCalcualtor) getDistance 메서드 함수 추가
